@@ -1,16 +1,13 @@
 {
   "name": "react-hardware",
-  "version": "0.5.0-alpha.2",
+  "version": "0.6.0-alpha.0",
   "description": "React firmata bindings.",
   "main": "lib/index.js",
   "jsnext:main": "src/index.js",
   "scripts": {
-    "install-prerelease": "npm install http://react.zpao.com/builds/master/latest/react.tgz",
     "flow": "flow",
     "lint": "eslint .",
-    "test": "./scripts/test",
-    "posttest": "npm run lint",
-    "postlint": "npm run flow",
+    "test": "jest",
     "prepublish": "npm run build",
     "build": "babel src -d lib",
     "example": "babel-node ./scripts/examples"
@@ -40,33 +37,37 @@
   },
   "homepage": "https://github.com/iamdustan/react-hardware",
   "dependencies": {
-    "fbjs": "^0.8.6",
-    "firmata": "^0.14.1",
-    "react": "16.0.0-alpha.2",
-    "react-devtools-core": "^2.0.12",
-    "react-dom": "16.0.0-alpha.2",
-    "serialport": "^4.0.7",
+    "firmata": "^2.0.0",
+    "invariant": "2.2.4",
+    "react": "16.8.4",
+    "react-devtools-core": "^3.6.1",
+    "react-reconciler": "^0.20.2",
+    "serialport": "^7.1.4",
     "setimmediate": "^1.0.5",
-    "ws": "^1.1.1"
+    "warning": "4.0.3"
   },
   "devDependencies": {
-    "babel-cli": "^6.6.4",
-    "babel-eslint": "^7.1.1",
-    "babel-plugin-transform-class-properties": "^6.6.0",
+    "@babel/cli": "^7.2.3",
+    "@babel/core": "^7.3.4",
+    "@babel/plugin-proposal-class-properties": "^7.3.4",
+    "@babel/preset-env": "^7.3.4",
+    "@babel/preset-flow": "^7.0.0",
+    "@babel/preset-react": "^7.0.0",
+    "@babel/register": "^7.0.0",
+    "babel-eslint": "^10.0.1",
     "babel-plugin-transform-react-jsx-source": "^6.22.0",
-    "babel-preset-es2015": "^6.6.0",
-    "babel-preset-react": "^6.5.0",
-    "babel-register": "^6.7.2",
-    "chokidar": "^1.3.0",
-    "commander": "^2.9.0",
-    "eslint": "^3.12.2",
-    "eslint-plugin-react": "^6.8.0",
-    "fbjs-scripts": "^0.7.1",
-    "flow-bin": "^0.38.0",
-    "jasmine": "^2.4.1",
-    "jasmine-es6": "0.1.8",
-    "jasmine-spec-reporter": "^2.4.0",
-    "mock-firmata": "0.0.7",
-    "react-devtools": "^2.0.11"
+    "commander": "^2.19.0",
+    "eslint": "^5.15.1",
+    "eslint-plugin-react": "^7.12.4",
+    "flow-bin": "^0.94.0",
+    "jest": "^24.1.0",
+    "prettier": "^1.16.4",
+    "mock-firmata": "0.2.0",
+    "react-devtools": "^3.6.1"
+  },
+  "prettier": {
+    "bracketSpacing": false,
+    "singleQuote": true,
+    "trailingComma": "all"
   }
 }
