@@ -1,15 +1,6 @@
 module.exports = {
   parser: 'babel-eslint',
-  'extends': './node_modules/fbjs-scripts/eslint/.eslintrc.js',
-  plugins: [
-    'react',
-  ],
-  globals: {
-    ReactComponent: true,
-    React$Component: true,
-    ReactClass: true,
-    React$Element: true,
-  },
+  plugins: ['react'],
   rules: {
     'object-curly-spacing': [2, 'never'],
     'react/jsx-uses-react': 2,
