@@ -3,16 +3,12 @@
  * This example blinks the Arduino Uno built-in LED on pin 13 every 1 second.
  */
 
-import React, {Component} from 'react';
+import * as React from 'react';
 import {getPort} from '../port';
 import ReactHardware from '../../src';
 
-class FlashingLed extends Component {
-  constructor(props, context) {
-    super(props, context);
-
-    this.state = {value: 1};
-  }
+class FlashingLed extends React.Component<{}, {value: number}> {
+  state = {value: 1};
 
   componentDidMount() {
     setInterval(() => {
@@ -23,20 +19,10 @@ class FlashingLed extends Component {
   }
 
   render() {
-    return (
-      <pin
-        pin={13}
-        value={this.state.value}
-        mode={'OUTPUT'}
-      />
-    );
+    return <pin pin={13} value={this.state.value} mode={'OUTPUT'} />;
   }
 }
 
-ReactHardware.render(
-  <FlashingLed />,
-  getPort(),
-  (inst) => {
-    console.log('Rendered <%s />', FlashingLed.name);
-  }
-);
+ReactHardware.render(<FlashingLed />, getPort(), inst => {
+  console.log('Rendered <%s />', FlashingLed.name);
+});
