@@ -18,29 +18,29 @@ import {getPort} from '../port';
 import ReactHardware from '../../src';
 
 type HardwareEvent = {
-  value: number;
-  type: string;
+  value: number,
+  type: string,
 };
 
 type P = {
-  pin: number;
-  onChange: ?(event:HardwareEvent) => any;
-  onDown: ?(event:HardwareEvent) => any;
-  onUp: ?(event:HardwareEvent) => any;
-}
+  pin: number,
+  onChange: ?(event: HardwareEvent) => any,
+  onDown: ?(event: HardwareEvent) => any,
+  onUp: ?(event: HardwareEvent) => any,
+};
 
 class Button extends Component {
   props: P;
   defaultProps: {};
-  onRead: (value:number) => any;
+  onRead: (value: number) => any;
 
-  constructor(props:P, context:{}) {
+  constructor(props: P, context: {}) {
     super(props, context);
 
     this.onRead = this.onRead.bind(this);
   }
 
-  onRead(value:number) {
+  onRead(value: number) {
     const {onDown, onUp, onChange} = this.props;
     if (value === 1 && typeof onDown === 'function') {
       onDown({value, type: 'down'});
@@ -54,13 +54,7 @@ class Button extends Component {
   }
 
   render() {
-    return (
-      <pin
-        pin={this.props.pin}
-        onRead={this.onRead}
-        mode={'INPUT'}
-      />
-    );
+    return <pin pin={this.props.pin} onRead={this.onRead} mode={'INPUT'} />;
   }
 }
 
@@ -71,7 +65,7 @@ class App extends Component {
     super();
 
     this.state = {on: false};
-    this.toggle = (event:HardwareEvent) => {
+    this.toggle = (event: HardwareEvent) => {
       this.setState({on: !this.state.on});
     };
   }
@@ -87,10 +81,6 @@ class App extends Component {
   }
 }
 
-ReactHardware.render(
-  <App />,
-  getPort(),
-  (inst) => {
-    console.log('Rendered <ButtonApplication />');
-  }
-);
+ReactHardware.render(<App />, getPort(), inst => {
+  console.log('Rendered <ButtonApplication />');
+});
