@@ -3,9 +3,7 @@
  * an environment variable or local, hard-coded value.
  */
 
-export const getPort = () => (
-  process.env.REACT_HARDWARE_EXAMPLE_PORT ||
-  process.env.NODE_ENV === 'test'
-    ? '/dev/cu.usbmodem1451' : null
-);
-
+export const getPort = () =>
+  process.env.REACT_HARDWARE_EXAMPLE_PORT || process.env.NODE_ENV === 'test'
+    ? '/dev/cu.usbmodem1451'
+    : null;
