@@ -5,16 +5,19 @@ const ws = new WebSocket('ws://localhost:8098');
 ws.onopen = () => console.log('Websocket connection opened');
 ws.onmessage = ({data}) => console.log(data);
 
-const SpeechRecognition = window.SpeechRecognition || window.webkitSpeechRecognition;
-const SpeechGrammarList = window.SpeechGrammarList || window.webkitSpeechGrammarList; // eslint-disable-line
-const SpeechRecognitionEvent = window.SpeechRecognitionEvent || window.webkitSpeechRecognitionEvent;
+const SpeechRecognition =
+  window.SpeechRecognition || window.webkitSpeechRecognition;
+const SpeechGrammarList =
+  window.SpeechGrammarList || window.webkitSpeechGrammarList; // eslint-disable-line
+const SpeechRecognitionEvent =
+  window.SpeechRecognitionEvent || window.webkitSpeechRecognitionEvent;
 
 const last = a => a[a.length - 1];
 
 type S = {
-  status: string;
-  hearing: string;
-  heard: Array<string>;
+  status: string,
+  hearing: string,
+  heard: Array<string>,
 };
 
 class App extends Component<any, any, S> {
@@ -33,7 +36,7 @@ class App extends Component<any, any, S> {
     this.onEnd = this.onEnd.bind(this);
   }
 
-  handleCommand(maybeCommand:string) {
+  handleCommand(maybeCommand: string) {
     if (/\bon\b/i.test(maybeCommand)) {
       ws.send('on');
     } else if (/\boff\b/i.test(maybeCommand)) {
@@ -50,7 +53,7 @@ class App extends Component<any, any, S> {
     this.recognition.continuous = true;
     this.recognition.interimResults = false;
     this.recognition.lang = 'en-US';
-    this.recognition.onerror = (event) => {
+    this.recognition.onerror = event => {
       if (event.error === 'no-speech') {
         this.setState({error: 'no-speech'});
       }
@@ -72,7 +75,7 @@ class App extends Component<any, any, S> {
     this.recognition = null;
   }
 
-  onClick(event:any) {
+  onClick(event: any) {
     if (this.state.status === 'LISTENING') {
       this.recognition.stop();
       this.setState({status: ''});
@@ -85,11 +88,11 @@ class App extends Component<any, any, S> {
     }
   }
 
-  onEnd(event:any) {
+  onEnd(event: any) {
     this.setState({status: '', hearing: ''});
   }
 
-  onResult(event:any) {
+  onResult(event: any) {
     [...event.results].reverse().some(r => {
       if (r.isFinal) {
         const maybeCommand = r[0].transcript;
@@ -105,10 +108,20 @@ class App extends Component<any, any, S> {
   render() {
     if (this.state.error) {
       return (
-        <div style={{background: 'yellow', padding: '0.75em', overflow: 'hidden'}}>
-          <strong style={{float: 'left', fontSize: '2em', marginRight: '0.5em'}}>:‘(</strong>
+        <div
+          style={{background: 'yellow', padding: '0.75em', overflow: 'hidden'}}
+        >
+          <strong
+            style={{float: 'left', fontSize: '2em', marginRight: '0.5em'}}
+          >
+            :‘(
+          </strong>
           <span>
-            It looks like your browser does not support the <a href="https://developer.mozilla.org/en-US/docs/Web/API/SpeechRecognition">SpeechRecognition</a> API.
+            It looks like your browser does not support the{' '}
+            <a href="https://developer.mozilla.org/en-US/docs/Web/API/SpeechRecognition">
+              SpeechRecognition
+            </a>{' '}
+            API.
           </span>
         </div>
       );
@@ -124,8 +137,12 @@ class App extends Component<any, any, S> {
           <span>{this.state.hearing}</span>
         </div>
         <div id="log">
-        <span>Everything heard: </span>
-          <ul>{this.state.heard.map((h, i) => <li key={i}>{h}</li>)}</ul>
+          <span>Everything heard: </span>
+          <ul>
+            {this.state.heard.map((h, i) => (
+              <li key={i}>{h}</li>
+            ))}
+          </ul>
         </div>
       </div>
     );
@@ -134,15 +151,20 @@ class App extends Component<any, any, S> {
 
 const Noapp = () => (
   <div style={{background: 'yellow', padding: '0.75em', overflow: 'hidden'}}>
-    <strong style={{float: 'left', fontSize: '2em', marginRight: '0.5em'}}>:‘(</strong>
+    <strong style={{float: 'left', fontSize: '2em', marginRight: '0.5em'}}>
+      :‘(
+    </strong>
     <span>
-      It looks like your browser does not support the <a href="https://developer.mozilla.org/en-US/docs/Web/API/SpeechRecognition">SpeechRecognition</a> API.
+      It looks like your browser does not support the{' '}
+      <a href="https://developer.mozilla.org/en-US/docs/Web/API/SpeechRecognition">
+        SpeechRecognition
+      </a>{' '}
+      API.
     </span>
   </div>
 );
 
 ReactDOM.render(
   SpeechRecognitionEvent ? <App /> : <Noapp />,
-  document.getElementById('app')
+  document.getElementById('app'),
 );
-
