@@ -9,44 +9,27 @@
  **/
 
 import type {HardwareEvent} from '../types';
-import React, {Component} from 'react';
+import * as React from 'react';
 
-type P = {
-  pin: number;
-  onChange: (event:HardwareEvent) => any;
-}
+type Props = {
+  pin: number,
+  onChange: (event: HardwareEvent) => any,
+};
 
-class Switch extends Component {
-  props: P;
-  static defaultProps: {};
-  value: number;
-  onRead: (value:number) => any;
+class Switch extends React.Component<Props> {
+  value = -1;
 
-  constructor(props:P, context:{}) {
-    super(props, context);
-
-    this.value = -1;
-    this.onRead = this.onRead.bind(this);
-  }
-
-  onRead(value:number) {
+  onRead = (value: number) => {
     const {onChange} = this.props;
     if (value !== this.value) {
       this.value = value;
       onChange({value, type: 'change'});
     }
-  }
+  };
 
   render() {
-    return (
-      <pin
-        pin={this.props.pin}
-        onRead={this.onRead}
-        mode={'INPUT'}
-      />
-    );
+    return <pin pin={this.props.pin} onRead={this.onRead} mode={'INPUT'} />;
   }
 }
 
 export default Switch;
-
