@@ -2,10 +2,10 @@
 
 const defineProperty = Object.defineProperty;
 defineProperty(global, 'WebSocket', {
-  value: require('ws')
+  value: require('ws'),
 });
 defineProperty(global, 'window', {
-  value: global
+  value: global,
 });
 
 const {connectToDevTools} = require('react-devtools-core');
@@ -19,4 +19,3 @@ connectToDevTools({
   // default port? port: ,
   resolveRNStyle: null, // TODO maybe: require('flattenStyle')
 });
-
