@@ -4,58 +4,34 @@
 
 import type {Board} from 'firmata';
 
-// import invariant from 'fbjs/lib/invariant';
-import warning from 'fbjs/lib/warning';
+// import invariant from 'invariant';
+import warning from 'warning';
 import {
   setPayloadForPin,
-  validatePayloadForPin
+  validatePayloadForPin,
 } from '../firmata/HardwareManager';
 
 const assertValidProps = validatePayloadForPin;
-
+import {create, diff} from '../attributePayload';
 type Instance = Object;
 type Props = Object;
-
-// TODO: switch `tag` to pin MODE
-/*
-const assertValidProps = (
-  tag : string,
-  rawProps : Object
-) => {
-  // TODO: getDeclarationErrorAddendum() and getCurrentOwnerName()
-  // should be added to the invariants
-  invariant(
-    rawProps.pin != null,
-    'A pin must be provided'
-  );
-
-  invariant(
-    rawProps.mode != null,
-    'A mode must be provided'
-  );
-
-  warning(
-    rawProps.value != null,
-    'A value should be provided'
-  );
-};
-*/
 
 const IO_KEY = '__IO__';
 
 const ReactHardwareFiberComponent = {
   createElement(
     tag: string,
-    props : Object,
-    rootContainerElement : Board,
-    hostContext : string | null
+    props: Object,
+    rootContainerElement: Board,
+    hostContext: string | null,
   ) {
     // Deprecated path for when a `container` is hit which was a hack in stack
     // for being unable to return an array from render.
     if (tag === 'container') {
-      warning(false,
+      warning(
+        false,
         'The <container /> tag has been deprecated and will be removed in ' +
-        'the next release. ReactFiber supports returning an array from render()'
+          'the next release. ReactFiber supports returning an array from render()',
       );
       return rootContainerElement;
     }
@@ -65,10 +41,10 @@ const ReactHardwareFiberComponent = {
   },
 
   setInitialProperties(
-    element : Instance,
-    tag : string,
-    rawProps : Object,
-    rootContainerElement : Board
+    element: Instance,
+    tag: string,
+    rawProps: Object,
+    rootContainerElement: Board,
   ) {
     // Deprecated path for when a `container` is hit which was a hack in stack
     // for being unable to return an array from render.
@@ -84,22 +60,22 @@ const ReactHardwareFiberComponent = {
   },
 
   diffProperties(
-    element : Instance,
-    tag : string,
-    lastRawProps : Object,
-    nextRawProps : Object,
-    rootContainerElement : Board
-  ) : void /*null | Array<mixed>*/ {
-
+    element: Instance,
+    tag: string,
+    lastRawProps: Object,
+    nextRawProps: Object,
+    rootContainerElement: Board,
+  ): void /*null | Array<mixed>*/ {
+    return nextRawProps;
   },
 
   updateProperties(
-    element : Instance,
-    updatePayload : Array<mixed>,
-    type : string,
-    oldProps : Props,
-    newProps : Props,
-    internalInstanceHandle : Object,
+    element: Instance,
+    updatePayload: Array<mixed>,
+    type: string,
+    oldProps: Props,
+    newProps: Props,
+    internalInstanceHandle: Object,
   ) {
     // Deprecated path for when a `container` is hit which was a hack in stack
     // for being unable to return an array from render.
@@ -116,4 +92,3 @@ const ReactHardwareFiberComponent = {
 };
 
 export default ReactHardwareFiberComponent;
-
