@@ -11,21 +11,20 @@ import {
   teardownConnection,
 } from './HardwareManager';
 
-const isBoardish = (thing) => (
+const isBoardish = thing =>
   thing.hasOwnProperty('id') &&
   thing.hasOwnProperty('repl') &&
-  thing.hasOwnProperty('pins')
-);
+  thing.hasOwnProperty('pins');
 
 function setup(port, callback) {
   if (isBoardish(port)) {
-    const board : Board = (port : any);
+    const board: Board = (port: any);
     console.warn(
       'It looks like you’re passing a Johnny-five board in. This is ' +
-      'experimental at best. Have fun! :)'
+        'experimental at best. Have fun! :)',
     );
     setupConnection('experimental-board', board);
-    board.on('ready', (error) => {
+    board.on('ready', error => {
       if (error) {
         callback(error);
       } else {
@@ -37,7 +36,7 @@ function setup(port, callback) {
   }
 
   console.info('Connecting to port "%s"', port);
-  const board = new Board(port, (error) => {
+  const board = new Board(port, error => {
     if (error) {
       console.info('Board setup error');
       callback(error);
@@ -48,7 +47,7 @@ function setup(port, callback) {
   });
   setupConnection(port, board);
 
-  board.on('error', (error) => {
+  board.on('error', error => {
     // TODO: look up docs/source code for this
     console.info('Board error event!');
     console.log(error);
@@ -59,18 +58,13 @@ function setup(port, callback) {
     // TODO: unmount React tree automatically when this happens.
     teardownConnection(port);
   });
-
 }
 
-type Result<T, E> =
-  | (error : E | null, value?: T) => void;
-  // | (error : null, root: T) => void;
+type Result<T, E> = (error: E | null, value?: T) => void;
+// | (error : null, root: T) => void;
 
 const HardwareInstanceManager = {
-  connect(
-    port : ?string,
-    callback: Result<Board, Error>
-  ) {
+  connect(port: ?string, callback: Result<Board, Error>) {
     if (port == null) {
       console.info('Requesting port...');
       Board.requestPort((error, port) => {
@@ -85,7 +79,7 @@ const HardwareInstanceManager = {
     }
   },
 
-  get(port : ?string) {
+  get(port: ?string) {
     if (port) {
       const connection = getConnection(port);
       if (connection) {
@@ -93,8 +87,7 @@ const HardwareInstanceManager = {
       }
     }
     return null;
-  }
+  },
 };
 
 export default HardwareInstanceManager;
-
