@@ -1,6 +1,6 @@
 /** @flow */
 import {Board} from 'firmata';
-import invariant from 'fbjs/lib/invariant';
+import invariant from 'invariant';
 import {analogToDigital} from './HardwarePinTranslations';
 
 /**
@@ -18,15 +18,15 @@ import {analogToDigital} from './HardwarePinTranslations';
  *   <pin mode="i2c" />
  */
 const FIRMATA_COMMUNICATION_METHOD = {
-  '0': 'digital',  // input
-  '1': 'digital',  // output
-  '2': 'analog',   // analog
-  '3': 'analog',   // pwm
-  '4': 'servo',    // servo
-  '5': 'UNKNOWN',  // shift
-  '6': 'i2c',      // i2c
-  '7': 'UNKNOWN',  // onewire
-  '8': 'UNKNOWN',  // stepper
+  '0': 'digital', // input
+  '1': 'digital', // output
+  '2': 'analog', // analog
+  '3': 'analog', // pwm
+  '4': 'servo', // servo
+  '5': 'UNKNOWN', // shift
+  '6': 'i2c', // i2c
+  '7': 'UNKNOWN', // onewire
+  '8': 'UNKNOWN', // stepper
   '16': 'UNKNOWN', // unknown
   '127': 'IGNORE', // ignore
 };
@@ -35,7 +35,7 @@ type TConnection<T, B> = {|
   status: T,
   rootID: string,
   board: B,
-  readers: {[pin:string]: {reader: Function, call: any}},
+  readers: {[pin: string]: {reader: Function, call: any}},
 |};
 
 export type Connection =
@@ -43,21 +43,18 @@ export type Connection =
   | TConnection<'CONNECTED', Board>
   | TConnection<'DISCONNECTED', null>;
 
-const deferredReader =
-  (connection, pin) =>
-    (value) => connection.readers[pin].call(value);
+const deferredReader = (connection, pin) => value =>
+  connection.readers[pin].call(value);
 
 const setReader = (
-  connection : Connection,
-  communicationType : string,
-  payload : Object
+  connection: Connection,
+  communicationType: string,
+  payload: Object,
 ) => {
-  if (typeof connection.setReader !== 'function') {
-    for (let c in connectionsByContainer) {
-      if (connectionsByContainer[c].board === connection) {
-        connection = connectionsByContainer[c];
-        break;
-      }
+  for (let c in connectionsByContainer) {
+    if (connectionsByContainer[c].board === connection) {
+      connection = connectionsByContainer[c];
+      break;
     }
   }
 
@@ -66,7 +63,10 @@ const setReader = (
     connection.readers[payload.pin] = {reader, call: null};
 
     // map A0-A5 to the appropriate analog index for node-firmata
-    const toNodeFirmataMapping = typeof payload.pin === 'string' ? parseInt(payload.pin.slice(1), 10) : payload.pin;
+    const toNodeFirmataMapping =
+      typeof payload.pin === 'string'
+        ? parseInt(payload.pin.slice(1), 10)
+        : payload.pin;
     /* $FlowFixMe computed property call */
     connection.board[`${communicationType}Read`](toNodeFirmataMapping, reader);
     connection.readers[payload.pin].call = payload.onRead;
@@ -75,13 +75,10 @@ const setReader = (
   connection.readers[payload.pin].call = payload.onRead;
 };
 
-const connectionsByContainer:{[key:string]: Connection} = {};
-export const getConnection = (port : string) => connectionsByContainer[port];
+const connectionsByContainer: {[key: string]: Connection} = {};
+export const getConnection = (port: string) => connectionsByContainer[port];
 
-export const setupConnection = (
-  port : string,
-  board : Board
-) => {
+export const setupConnection = (port: string, board: Board) => {
   const connection: TConnection<'CONNECTING', Board> = {
     rootID: port,
     status: 'CONNECTING',
@@ -93,45 +90,44 @@ export const setupConnection = (
 };
 
 export const updateConnection = (
-  port : string,
-  board : Board
+  port: string,
+  board: Board,
 ): TConnection<'CONNECTED', Board> => {
-  const connection = connectionsByContainer[port];
-  if (!connection) {
-    throw new Error('Attempted to update non-existent connection for port ' + port);
+  const conn = connectionsByContainer[port];
+  if (!conn) {
+    throw new Error(
+      'Attempted to update non-existent connection for port ' + port,
+    );
   }
-  return {
-    rootID: connection.rootID,
-    readers: connection.readers,
-    board: board,
-    status: 'CONNECTED'
-  };
+  const connection: TConnection<'CONNECTED', Board> = (conn: any);
+  connection.status = 'CONNECTED';
+  return connection;
 };
 
-export const teardownConnection = (port : string): null | TConnection<'DISCONNECTED', null> => {
-  const connection = connectionsByContainer[port];
-  if (!connection) {
+export const teardownConnection = (
+  port: string,
+): null | TConnection<'DISCONNECTED', null> => {
+  const conn = connectionsByContainer[port];
+  if (!conn) {
     console.warn(
-      'Attempted to teardown non-existent connection for port %s', 
-      port
+      'Attempted to teardown non-existent connection for port %s',
+      port,
     );
     return null;
   } else {
+    const connection: TConnection<'DISCONNECTED', null> = (conn: any);
     // TODO : memory leak. Remove these
     for (const reader in connection.readers) {
       delete connection.readers[reader];
     }
-    return {
-      status: 'DISCONNECTED',
-      board: null,
-      rootID: connection.rootID,
-      readers: connection.readers
-    };
+    connection.status === 'DISCONNECTED';
+    connection.board = null;
+    return connection;
   }
 };
 
 // matches return value of the input value
-const findConnectionForRootId = (rootID) : Connection | null => {
+const findConnectionForRootId = (rootID): Connection | null => {
   for (const connection in connectionsByContainer) {
     if (connectionsByContainer[connection].rootID !== rootID) {
       continue;
@@ -148,24 +144,25 @@ const findConnectionForRootId = (rootID) : Connection | null => {
  * configuration.
  */
 export const validatePayloadForPin = (
-  maybeConnection : string | Board,
-  payload : Object
+  maybeConnection: string | Board,
+  payload: Object,
 ) => {
   if (payload == null) {
     return;
   }
 
-  const connection = typeof maybeConnection === 'string'
-    ? findConnectionForRootId(maybeConnection)
-    : maybeConnection;
+  const connection =
+    typeof maybeConnection === 'string'
+      ? findConnectionForRootId(maybeConnection)
+      : maybeConnection;
 
   invariant(
     !!connection,
     'Attempting to update connection string "%s" that no longer exists',
-    connection
+    connection,
   );
 
-  const board : Board = (connection.board || connection : any);
+  const board: Board = (connection.board || connection: any);
   const {pins, MODES} = board;
 
   const mode = MODES[payload.mode];
@@ -180,9 +177,11 @@ export const validatePayloadForPin = (
   const {supportedModes, analogChannel} = pins[normalizedPin];
   invariant(
     supportedModes.indexOf(mode) !== -1 ||
-    (analogChannel === 127 && payload.mode === 'DIGITAL'),
+      (analogChannel === 127 && payload.mode === 'DIGITAL'),
     'Unsupported mode "%s" for pin "%s".\nSupported modes are: "%s"',
-    payload.mode, payload.pin, supportedModes.map(m => idToModeName[m]).join('", "') || 'DIGITAL'
+    payload.mode,
+    payload.pin,
+    supportedModes.map(m => idToModeName[m]).join('", "') || 'DIGITAL',
   );
 };
 
@@ -190,37 +189,36 @@ export const validatePayloadForPin = (
  * Sets a pin's values to the desired payload.
  */
 export const setPayloadForPin = (
-  maybeConnection : string | Connection | Board,
-  payload : ?Object
+  maybeConnection: string | Connection | Board,
+  payload: ?Object,
 ) => {
   if (payload == null) {
     return;
   }
 
-  const connection = typeof maybeConnection === 'string' ?
-    findConnectionForRootId(maybeConnection) :
-    maybeConnection;
+  const connection =
+    typeof maybeConnection === 'string'
+      ? findConnectionForRootId(maybeConnection)
+      : maybeConnection;
 
   if (!connection) {
     return;
   }
 
   // backwards compatible with Stack
-  const board : Board = (connection.board || connection : any);
+  const board: Board = (connection.board || connection: any);
   const {MODES} = board;
 
-  // console.log(`set pinMode of "%s" to "%s"`, payload.pin, payload.mode);
   const normalizedPin = analogToDigital(payload.pin);
   board.pinMode(normalizedPin, MODES[payload.mode]);
   const communicationType = FIRMATA_COMMUNICATION_METHOD[MODES[payload.mode]];
   if (typeof payload.value !== 'undefined') {
-    // console.log(`${communicationType}Write to "%s" with "%s"`, payload.pin, payload.value);
     /* $FlowFixMe computed property call */
     board[`${communicationType}Write`](payload.pin, +payload.value);
   }
 
   if (payload.onRead) {
-    setReader((connection : any), communicationType, payload);
+    setReader((connection: any), communicationType, payload);
   }
 };
 
@@ -228,8 +226,8 @@ export const setPayloadForPin = (
  * NOTE: This is a leaky abstraction. It returns the direct Board IO instance.
  */
 export const getNativeNode = (
-  component : ReactComponent<*, *, *> & {_rootNodeID:string}
-) : Board | null => {
+  component: ReactComponent<*, *, *> & {_rootNodeID: string},
+): Board | null => {
   const connection = findConnectionForRootId(component._rootNodeID);
 
   if (connection) {
@@ -238,4 +236,3 @@ export const getNativeNode = (
     return null;
   }
 };
-
